refactor(web): settings, request, and integration components (#1062)
* remove styles

* remove commented code from RequestStatus

* remove commented code from RequestList/hooks.ts

* move the link component from molecules into organisms level

* remove commented code from molecules/MyIntegrations/Conten
