+++ web/src/components/organisms/Project/Request/RequestList/index.tsx
@@ -12,8 +12,6 @@ const RequestList: React.FC = () => {
     loading,
     collapsedCommentsPanel,
     selectedRequest,
-    // selectRequest,
-    // selectedRequests,
     selection,
     setSelection,
     collapseCommentsPanel,
